Clean up tablet header validation
Summary:
This collects the header validation checks into their own function and
makes the type of BlockEncoding explicitly a uint8.

Test Plan: go test

Reviewers: tberman, jyeo, jimmy

Reviewed By: jyeo

CC: justin

Differential Revision: http://phabricator.thefactory.com/D626

+++ go/datastore/tablet_writer.go
@@ -153,7 +153,7 @@ func writeBinary(w io.Writer, data interface{}) (uint32, error) {
 }
 
 func writeHeader(w io.Writer, opts *TabletOptions) uint32 {
-	h := Header{tabletMagic, uint8(opts.BlockEncoding), 0, 0, 0}
+	h := Header{tabletMagic, opts.BlockEncoding, 0, 0, 0}
 	n, _ := writeBinary(w, &h)
 	return n
 }
